refactor(test): split ci benchmark harness
